% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_age_mu_devs.R
\name{compute_age_mu_devs}
\alias{compute_age_mu_devs}
\title{Compute Age-Specific Movement Deviations and Baseline Movement Rates}
\usage{
compute_age_mu_devs(move_matrix, n_stocks, n_seasons)
}
\arguments{
\item{move_matrix}{A 3D array of movement rates with dimensions `(n_regions, n_regions-1, n_ages)`.
Each entry represents the movement rate from one region to another across different ages.}

\item{n_stocks}{An integer specifying the number of stocks for which deviations should be computed.}
}
\value{
A list containing:
  \describe{
    \item{`age_mu_devs`}{A 4D array of movement deviations with dimensions `(n_stocks, n_regions, n_regions-1, n_ages)`,
                          ensuring all values are non-negative.}
    \item{`baseline_move`}{A 3D array `(n_regions, n_regions-1, n_ages)` representing the minimum movement
                          rate for each region-pair across ages, used as a baseline.}
    \item{`mean_vals`}{A 4D array `(n_stocks, n_seasons, n_regions, n_regions-1)` representing the minimum movement
                          rate for each region-pair across stocks and season, used as a baseline mean movement rate.}
  }
}
\description{
This function calculates the deviation of movement rates at age from the minimum movement rate
for each region-to-region pair, ensuring that all deviations are greater than or equal to zero.
The function also saves the baseline movement rates, which can be used to reconstruct the
original movement matrix.
}
