% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_generate_em_info.R
\name{filter_and_generate_em_info}
\alias{filter_and_generate_em_info}
\title{Filter Fleets and Indices for an Estimation Model}
\usage{
filter_and_generate_em_info(
  em_info,
  fleet_regions,
  index_regions,
  filter_indices = NULL,
  em.opt
)
}
\arguments{
\item{em_info}{List. The original estimation model input data.}

\item{fleet_regions}{Integer vector. Specifies the region assignment of each fleet.}

\item{index_regions}{Integer vector. Specifies the region assignment of each index.}

\item{filter_indices}{Integer vector (optional). Indicates which indices to keep (1) or exclude (0).}
}
\value{
A list containing modified `em_info` lists for each region with filtered fleets, indices, renumbered index regions, and extracted weight-at-age (WAA) matrices.
}
\description{
This function filters fleets and indices based on region specifications
and removes indices marked for exclusion while ensuring correct renumbering.
Additionally, it extracts and correctly assembles the weight-at-age (WAA) matrix
for each region based on fleet, region, index, and stock assignments.
}
