% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_move.R
\name{fix_move}
\alias{fix_move}
\title{Fix movement parameter}
\usage{
fix_move(input)
}
\arguments{
\item{input}{wham input}
}
\value{
a wham input with movement rate treated as known
}
\description{
a function to fix movement rate in the input.
}
\examples{
\dontrun{
basic_info <- generate_basic_info()
move <- generate_move(basic_info = basic_info, move.type = 2, move.rate = 0.3, move.re = "constant")
input <- prepare_wham_input(basic_info = basic_info, move = move)
input <- fix_move(input)
}

}
