% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_info_em_SEP.R
\name{generate_basic_info_em_SEP}
\alias{generate_basic_info_em_SEP}
\title{Generate Basic Information for Separate Independent Estimation Models}
\usage{
generate_basic_info_em_SEP(
  em_info,
  em_years,
  n_stocks = NULL,
  n_regions = NULL,
  n_fleets = NULL,
  n_indices = NULL,
  filter_indices = NULL
)
}
\arguments{
\item{em_info}{List. Contains parameter inputs from a previously defined estimation model.}

\item{em_years}{Numeric vector. Years to be used for the estimation model.}

\item{n_stocks}{Integer. (Optional) Number of stocks; defaults to `em_info$par_inputs$n_stocks`.}

\item{n_regions}{Integer. (Optional) Number of regions; defaults to `em_info$par_inputs$n_regions`.}

\item{n_fleets}{Integer. (Optional) Number of fleets; defaults to `em_info$par_inputs$n_fleets`.}

\item{n_indices}{Integer. (Optional) Number of indices; defaults to `em_info$par_inputs$n_indices`.}

\item{filter_indices}{Integer (0/1) vector (optional). User-specified which indices are excluded from the assessment model}

\item{n_seasons}{Integer. (Optional) Number of seasons; defaults to `em_info$par_inputs$n_seasons`.}
}
\value{
A list containing the generated biological and fishery information,
        formatted for input into WHAM or other stock assessment models.
}
\description{
This helper function generates basic biological and fishery information
for simulation-estimation and management strategy evaluation (MSE).
It builds on `generate_basic_info` by allowing parameter values
to be inherited from an existing estimation model (EM) setup.
}
\seealso{
`generate_basic_info`, `prepare_wham_input`
}
