% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Update the operating model and generate data}
\usage{
generate_data(om, seed = 123)
}
\arguments{
\item{om}{Operating model}

\item{seed}{Seed used to generate data}
}
\value{
an operating model with simulated data
}
\description{
Function to use the operating model to generate data
}
\examples{
\dontrun{
basic_info <- generate_basic_info()
input <- prepare_wham_input(basic_info = basic_info)
mod <- fit_wham(input, do.fit = FALSE)
data <- generate_data(mod, seed = 123)
}

}
