% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ini_N1.R
\name{generate_ini_N1}
\alias{generate_ini_N1}
\title{Generate initial numbers-at-age}
\usage{
generate_ini_N1(log_N1 = c(10.6, 10), basic_info, ini.opt = "equilibrium")
}
\arguments{
\item{log_N1}{a vector (length = n_stocks) of log numbers-at-age1}

\item{basic_info}{(optional) list specifying options for numbers-at-age random effects, initial parameter values, and recruitment model (see details)}

\item{ini.opt}{N1_model
\describe{
   \item{"age-specific-fe"}{(default) age- and region-specific fixed effects parameters}
   \item{"equilibrium"}{2 fixed effects parameters: an initial recruitment and an instantaneous fishing mortality rate to generate an equilibruim abundance at age.}
   \item{"iid-re"}{(default) age- and region-specific iid random effects parameters. 2 parameters: mean and sd for log NAA}
   \item{"ar1-re"}{(default) age- and region-specific random effects parameters. 3 parameters: mean and sd, and cor for log NAA}
}}
}
\value{
an array of numbers-at-age (n_stocks x n_regions x n_ages)
}
\description{
Generate initial numbers-at-age
}
\examples{
\dontrun{
basic_info <- generate_basic_info()
N1_pars <- generate_ini_N1(log_N1 = c(10.6,10), basic_info, ini.opt = "equilibrium") 
}

}
