% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_move.R
\name{generate_move}
\alias{generate_move}
\title{Configure movement random effects}
\usage{
generate_move(
  basic_info,
  move.type = 2,
  move.rate = c(0.3, 0.2),
  move.re = "constant",
  move.sigma = 0.5,
  move.rho_a = 0.5,
  move.rho_y = 0.5
)
}
\arguments{
\item{basic_info}{Information generated by \code{\link{generate_basic_info}}.
Note: NAA_where has to be specified in basic_info using \code{\link{generate_NAA_where}}.}

\item{move.type}{Movement type
\itemize{
  \item \code{1} unidirectional movement
  \item \code{2} bidirectional movement (default)
  \item \code{3} no movement
}}

\item{move.rate}{Movement rate (default = c(0.3, 0.2))}

\item{move.re}{Movement random effects
\itemize{
  \item \code{"constant"} constant movement rate across years and ages (default).
  \item \code{"iid_y"} movement rate deviations vary by year, but uncorrelated.
  \item \code{"iid_a"} movement rate deviations vary by age, but uncorrelated.
  \item \code{"ar1_y"} movement rate deviations correlated by year (AR1).
  \item \code{"ar1_a"} movement rate deviations correlated by age (AR1).
}}

\item{move.sigma}{Sigma for movement random effects (default = 0.5)}

\item{move.rho_a}{Correlation for movement random effects by ages, only used if move.re = "ar1_a" (default = 0.5)}

\item{move.rho_y}{Correlation for movement random effects by years, only used if move.re = "ar1_y" (default = 0.5)}
}
\value{
A named list with the following components:
  \describe{
    \item{$stock_move}{Length = n_stocks, TRUE/FALSE whether each stock can move. If not provided, movement will be defined below for all stocks.}
    \item{$separable}{Length = n_stocks, TRUE/FALSE whether movement should be modeled separably from mortality or both occurring simultaneously.}
    \item{$must_move}{Array (n_stocks x n_seasons x n_regions) 0/1 indicator whether movement from a region must occur.}
    \item{$can_move}{Array (n_stocks x n_seasons x n_regions x n_regions) 0/1 indicator whether movement can occur from one region to another.}
  }
}
\description{
The generate_move() function is used to configure movement random effects.
}
\examples{
\dontrun{
data <- generate_basic_info()
data <- generate_move(data, move.type = 2, move.rate = c(0.3,0.1), move.re = "constant")
}
}
\seealso{
\code{\link{set_move}} and \code{\link{generate_NAA_where}}
}
