% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_NAA_where.R
\name{generate_NAA_where}
\alias{generate_NAA_where}
\title{Specify movement type}
\usage{
generate_NAA_where(basic_info = basic_info, move.type = 2)
}
\arguments{
\item{basic_info}{Information generated by \code{\link{generate_basic_info}}}

\item{move.type}{Movement type:
\itemize{
  \item \code{1} single stock movement (i.e. only stock 1 can move), with region-specific movement rates
  \item \code{2} bidirectional movement (i.e. all stocks can move), with region-specific movement rates (default)
  \item \code{3} no movement
  }}
}
\value{
A list of information that will be used to prepare wham input
}
\description{
The generate_NAA_where() function is used to generate an array (n_stocks,n_regions,n_ages) 
which determines whether or not (T/F) the stock from a certain age can be present in a specific region.
}
\examples{
\dontrun{
basic_info <- generate_basic_info()
basic_info <- generate_NAA_where(basic_info = basic_info, move.type = 2)
}

}
\seealso{
\code{\link{prepare_wham_input}} and \code{\link{generate_move}}
}
