% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_F_from_Catch_region.R
\name{get_F_from_Catch_region}
\alias{get_F_from_Catch_region}
\title{Calculate region-specific F from region-specific catch}
\usage{
get_F_from_Catch_region(
  Catch,
  NAA,
  log_M,
  mu,
  L,
  sel,
  fracyr_season,
  fleet_regions,
  fleet_seasons,
  can_move,
  mig_type,
  waacatch,
  trace,
  F_init
)
}
\arguments{
\item{Catch}{Vector of catch values for each region.}

\item{NAA}{Array (n_stocks x n_regions x ny x n_ages); numbers-at-age.}

\item{log_M}{Array (n_stocks x n_regions x ny x n_ages); natural mortality.}

\item{mu}{Array (n_stocks x n_ages x n_seasons x n_years x n_regions x n_regions); movement rates.}

\item{L}{Matrix (n_years x n_regions); "extra" mortality rate.}

\item{sel}{Matrix of selectivity values.}

\item{fracyr_season}{Vector of length n_seasons; length of intervals for each season.}

\item{fleet_regions}{Vector of length n_fleets; indicates which region each fleet is operating in.}

\item{fleet_seasons}{Matrix (n_fleets x n_seasons); 0/1 indicating whether fleet is operating in the season.}

\item{can_move}{Array (n_stocks x n_seasons x n_regions x n_regions); 0/1 indicating whether movement can occur from one region to another.}

\item{mig_type}{Vector of length n_stocks; 0 = migration after survival, 1 = movement and mortality simultaneous.}

\item{waacatch}{Matrix (n_fleets x n_ages); weight-at-age.}

\item{trace}{Logical; whether to print information to the screen.}

\item{F_init}{Vector of initial F values for each region.}
}
\value{
A vector of region-specific F values estimated using Newton's method.
}
\description{
The get_F_from_Catch_region() function calculates region-specific F given the aggregate region-specific catch.
Note that some helper functions that support the get_F_from_Catch_region() function are included below.
}
