% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_input_no_asap_fn.r
\name{initial_input_no_asap_fn}
\alias{initial_input_no_asap_fn}
\title{Prepare Input Without ASAP File}
\usage{
initial_input_no_asap_fn(input, basic_info)
}
\arguments{
\item{input}{Initial wham input}

\item{basic_info}{A list of basic information}
}
\value{
wham input file
}
\description{
Function to prepare wham input without ASAP file
}
