% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_through_fn.R
\name{loop_through_fn}
\alias{loop_through_fn}
\title{Function to perform management strategy evaluation}
\usage{
loop_through_fn(
  om,
  random,
  M_om,
  sel_om,
  NAA_re_om,
  move_om,
  age_comp_om = "multinomial",
  M_em,
  sel_em,
  NAA_re_em,
  move_em,
  em.opt = NULL,
  age_comp_em = "multinomial",
  assess_years = NULL,
  assess_interval = NULL,
  base_years = NULL,
  year.use = 30,
  hcr.type = 1,
  hcr.opts = NULL,
  do.retro = FALSE,
  do.osa = FALSE,
  seed = 123,
  save.sdrep = FALSE
)
}
\arguments{
\item{om}{Operating model including pseudo data}

\item{random}{A vector of processes that are treated as random effects in the operating model}

\item{M_om}{Natural mortality configuration in the operating model}

\item{sel_om}{Selectivity configuration in the operating model}

\item{NAA_re_om}{Numbers-at-age (NAA) configuration in the operating model}

\item{move_om}{Movement configuration in the operating model}

\item{age_comp_om}{Likelihood distribution of age composition data in the operating model}

\item{M_em}{Natural mortality configuration in the assessment model}

\item{sel_em}{Selectivity configuration in the assessment model}

\item{NAA_re_em}{Numbers-at-age (NAA) configuration in the assessment model}

\item{move_em}{Movement configuration in the assessment model}

\item{em.opt}{List of options for the assessment model}

\item{age_comp_em}{Likelihood distribution of age composition data in the assessment model}

\item{assess_years}{Year in which the assessment is conducted}

\item{assess_interval}{Assessment interval used in the MSE feedback loop}

\item{base_years}{Years used in the burn-in period}

\item{year.use}{Number of years used in the assessment model}

\item{hcr.type}{Type of harvest control rule}

\item{hcr.opts}{A list of HCR information, only used if hcr.type = 3}

\item{do.retro}{T/F Do retrospective analysis? Default = TRUE}

\item{do.osa}{T/F Calculate one-step-ahead (OSA) residuals? Default = TRUE}

\item{seed}{Seed used for generating data}

\item{save.sdrep}{T/F Save every assessment model (memory intensive)
\itemize{
  \item \code{"TRUE"} Save every assessment model 
  \item \code{"FALSE"} Save the last assessment model (default)
  }}
}
\value{
a list of model output
}
\description{
A wrapper function to step through the feedback period, update the operating model,
refit the estimation model, and generate catch advice.
}
\seealso{
\code{\link{make_em_input}}, \code{\link{update_om_fn}}, \code{\link{advice_fn}}
}
