% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregate_data.R
\name{make_aggregate_data}
\alias{make_aggregate_data}
\title{Aggregate Data for Estimation Model}
\usage{
make_aggregate_data(
  om,
  em_info,
  aggregate_catch_info,
  aggregate_index_info,
  ind_em
)
}
\arguments{
\item{om}{List. The operating model containing observed data, including:}

\item{em_info}{List. The estimation model information, including:}

\item{aggregate_catch_info}{List. Contains details for aggregating fleet-specific catch data:
\itemize{
  \item `fleet_pointer` - Integer vector of length `n_fleets`, specifying how fleets should be aggregated.
  \item `catch_cv`, `catch_Neff` - Numeric vectors of length `n_fleets`, specifying catch CV and sample size.
  \item `use_agg_catch`, `use_catch_paa` - Boolean vectors of length `n_fleets`, indicating whether to aggregate.
}}

\item{aggregate_index_info}{List. Contains details for aggregating index-specific data:
\itemize{
  \item `index_pointer` - Integer vector of length `n_indices`, specifying how indices should be aggregated.
  \item `index_cv`, `index_Neff`, `q` - Numeric vectors of length `n_indices`, specifying index CV, effective sample size, and catchability coefficient.
  \item `use_indices`, `use_index_paa` - Boolean vectors of length `n_indices`, indicating whether to aggregate.
}}

\item{ind_em}{Vector. Indices specifying the years for which the estimation model should use data.}
}
\value{
List. An updated `em_info` object with:
  \itemize{
    \item Aggregated `catch_info` (`agg_catch`, `catch_paa`).
    \item Aggregated `index_info` (`agg_indices`, `index_paa`).
    \item Updated `par_inputs` with fleet/index aggregations.
    \item Updated weight-at-age (`user_waa`) and `basic_info$waa`.
  }
}
\description{
Aggregates catch and index data from the operating model (`om`) and updates
the estimation model (`em_info`) with the aggregated data. It also adjusts
`par_inputs` and weight-at-age (`user_waa`) to reflect the modified fleet,
region, index, and stock structure.
}
\details{
- Fleets and indices with `fleet_pointer = 0` or `index_pointer = 0` are ignored during aggregation.
- **Aggregated parameters** (`catch_cv`, `index_cv`, `q`, etc.) will return **vectors of length equal to the number of aggregated fleets or indices**.
- If `aggregate_catch_info` or `aggregate_index_info` are missing values, they are replaced with the first element of `em_info$par_inputs`.
- Weight-at-age (`waa`) is aggregated based on fleet or index weights, using either **catch-weighted averages or equal proportions**.
}
\examples{
\dontrun{
# Example usage
new_em_info <- make_aggregate_data(om, em_info, aggregate_catch_info, aggregate_index_info, ind_em)
}

}
