% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_em_input.R
\name{make_em_input}
\alias{make_em_input}
\title{Generate input data for the assessment model}
\usage{
make_em_input(
  om,
  em_info,
  M_em,
  sel_em,
  NAA_re_em,
  move_em,
  em.opt,
  em_years,
  year.use,
  age_comp_em = "multinomial",
  aggregate_catch_info = NULL,
  aggregate_index_info = NULL
)
}
\arguments{
\item{om}{Operating model}

\item{em_info}{A list of information used to generate the operating model}

\item{M_em}{Natural mortality random effects}

\item{sel_em}{Selectivity random effects}

\item{NAA_re_em}{Numbers-at-age random effects}

\item{move_em}{Movement random effects}

\item{em.opt}{Movement random effects options
\itemize{
  \item \code{$separate.em} TRUE = No Global SPR, FALSE = Global SPR
  \item \code{$separate.em.type} only if separate.em = TRUE \cr
  {=1} panmictic (spatially-aggregated) \cr
  {=2} fleets-as-areas \cr
  {=3} n single assessment models (n = n_regions) \cr
  \item \code{$do.move} T/F movement is included (use if separate.em = FALSE)
  \item \code{$est.move} T/F movement rate is estimated (use if separate.em = FALSE)
}}

\item{em_years}{Years used in the assessment model}

\item{year.use}{Number of years used in the assessment model}

\item{age_comp_em}{Likelihood distribution of age composition data in the assessment model
\itemize{
  \item \code{"multinomial"} Default
  \item \code{"dir-mult"}
  \item \code{"dirichlet-miss0"}
  \item \code{"dirichlet-pool0"}
  \item \code{"logistic-normal-miss0"}
  \item \code{"logistic-normal-ar1-miss0"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl"}
  \item \code{"logistic-normal-01-infl-2par"}
  \item \code{"mvtweedie"}
  \item \code{"dir-mult-linear"}
}}

\item{aggregate_catch_info}{(optional) User specified list of information for aggregate catch (default: use first fleet). Only when panmictic model with aggregate catch is used.
\itemize{
  \item \code{$catch_cv} {vector (n_fleets) of CVs for fleet catch.}
  \item \code{$catch_Neff} {vector (n_fleets) of effective sample sizes for fleet catch.}
  \item \code{$use_agg_catch} {vector (n_fleets) of 0/1 values flagging whether whether to use aggregate catch.}
  \item \code{$use_catch_paa} {vector (n_fleets) of 0/1 values flagging whether to use proportions at age observations.}
  \item \code{$fleet_pointer} {vector (n_fleets) of which fleets should be combined.}
  }}

\item{aggregate_index_info}{(optional) User specified list of information for aggregate index (default: use first index). Only when panmictic model with aggregate index is used.
\itemize{
  \item \code{$index_cv} {vector (n_indices) of CVs for survey indices.}
  \item \code{$index_Neff} {vector (n_indices) of effective sample sizes for survey indices.}
  \item \code{$fracyr_indices} {vector (n_indices) of fractions of the year when each survey is conducted.}
  \item \code{$q} {vector (n_indices) of survey catchabilities.}
  \item \code{$use_indices} {vector (n_indices) of 0/1 values flagging whether to use aggregate observations.}
  \item \code{$use_index_paa} {vector (n_indices) of 0/1 values flagging whether to use proportions at age observations.}
  \item \code{$units_indices} {vector (n_indices) of 1/2 values flagging whether aggregate observations are biomass (1) or numbers (2).}
  \item \code{$units_index_paa} {vector (n_indices) of 1/2 values flagging whether composition observations are biomass (1) or numbers (2).}
  \item \code{$index_pointer} {vector (n_indices) of which indices should be combined.}
  }}
}
\value{
a wham input
}
\description{
A function to generate the input for the estimation model for management strategy evaluation.
}
\examples{
\dontrun{
data <- generate_basic_info()
input <- prepare_wham_input(basic_info = data)
mod <- fit_wham(input, do.fit = FALSE)
data <- make_em_input(mod, em.opt = list(separate.em = TRUE, separate.em.type = 1, do.move = FALSE, est.move = FALSE), year.use = 10, em_years = 1973:2022)
}
}
\seealso{
\code{\link{loop_through_fn}}
}
