% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mse_output.R
\name{plot_mse_output}
\alias{plot_mse_output}
\title{Generate MSE Output Plots and Report}
\usage{
plot_mse_output(
  mods,
  main.dir = getwd(),
  use.n.years = 10,
  base.mod = 1,
  short.term = NULL,
  dpi = 150,
  out.type = "png"
)
}
\arguments{
\item{mods}{List of models to compare.}

\item{main.dir}{Main directory for output files. Default is the current working directory.}

\item{use.n.years}{Number of years to use for comparison. Default is 10.}

\item{base.mod}{Base model for comparison. Default is 1.}

\item{short.term}{List with first.n.years, year_start, and year_end for short-term analysis. if = NULL, results will be summarized for the first 2 years in the feedback loop}

\item{dpi}{Resolution for plots. Default is 150.}

\item{out.type}{Output type for plots.
\itemize{
  \item \code{"png"} median-lived species (default)
  \item \code{"pdf"} short-lived species
  \item \code{"html"} long-lived species
  }}
}
\value{
Generates plots in a report.
}
\description{
This function generates plots for MSE output and creates a report in png/pdf/HTML format.
}
\examples{
plot_mse_output(mods, main.dir = "your/directory/path", out.type = 'pdf')

}
