% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_F.R
\name{update_om_F}
\alias{update_om_F}
\title{Update fishing mortality in the operating model given the catch}
\usage{
update_om_F(om, year, catch)
}
\arguments{
\item{om}{Operating model}

\item{year}{Actual year(s) in the projection}

\item{catch}{Catch advice in the projection years}
}
\value{
an operating model with updated F time series
}
\description{
Function to update F in the operating model given the catch advice
}
\seealso{
\code{\link{get_F_from_catch}}
}
