% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_fn.R
\name{advice_fn}
\alias{advice_fn}
\title{Generate Catch Advice for Management Strategy Evaluation}
\usage{
advice_fn(em, pro.yr = assess.interval, hcr)
}
\arguments{
\item{em}{An estimation model fitted using WHAM.}

\item{pro.yr}{Integer. The number of years for projection.}

\item{hcr}{A list specifying the harvest control rule (HCR) type and options. It should contain:
\describe{
  \item{\code{hcr.type}}{Integer. The type of harvest control rule to use. Options include:}
  \itemize{
    \item \code{1}: Annual projected catch based on 75\% of F40\% (default).
    \item \code{2}: Constant catch based on 75\% of F40\%.
    \item \code{3}: "Hockey stick" approach based on stock status.
  }
  \item{\code{hcr.opts}}{A list of additional options for the harvest control rule, including:}
  \describe{
    \item{\code{percentFXSPR}}{Numeric. The percentage of F\_XSPR to use for calculating catch in projections (default = 75).}
    \item{\code{percentSPR}}{Numeric. The X\% SPR used to calculate reference points (default = 40).}
    \item{\code{avg.yrs}}{Integer. The number of years over which mean values of MAA, WAA, M, and F will be averaged in projections (default = 5).}
    \item{\code{max_percent}}{Numeric. Maximum percentage of F\_XSPR to use in projections (default = 75).}
    \item{\code{min_percent}}{Numeric. Minimum percentage of F\_XSPR to use in projections (default = 0.01).}
    \item{\code{BThresh_up}}{Numeric. Upper bound of overfished biomass threshold (default = 0.5).}
    \item{\code{BThresh_low}}{Numeric. Lower bound of overfished biomass threshold (default = 0.1).}
  }
}}
}
\value{
A matrix containing the projected catch advice for the specified number of projection years.
}
\description{
A function to generate catch advice based on different harvest control rules (HCRs) for use in management strategy evaluation.
}
\examples{
\dontrun{
data <- generate_basic_info(n_stocks = 2, n_regions = 2, n_indices = 2, n_fleets = 2, base.years = 2003:2022)
NAA_re <- list(N1_model = c("equilibrium", "equilibrium"), sigma = c("rec", "rec"), cor = c("iid", "iid"), recruit_model = 2)
input <- prepare_wham_input(basic_info = data, NAA_re = NAA_re)
mod <- fit_wham(input, do.fit = FALSE)
input$data <- mod$simulate(complete = TRUE)
mod <- fit_wham(input, do.osa = FALSE, do.retro = FALSE, MakeADFun.silent = FALSE)
catch <- advice_fn(mod, pro.yr = 3, hcr = list(hcr.type = 1))
}
}
\seealso{
\code{\link{project_wham}}
}
