% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_fn.R
\name{advice_fn}
\alias{advice_fn}
\title{Generate catch advice}
\usage{
advice_fn(em, pro.yr = assess.interval, hcr.type = 1, hcr.opts = NULL)
}
\arguments{
\item{em}{Estimation model}

\item{pro.yr}{Number of years in projection}

\item{hcr.type}{Type of harvest control rule
\itemize{
  \item \code{"1"} Annual projected catch based on 75% of F40% (default)
  \item \code{"2"} Constant catch based on 75% of F40% (default)
  \item \code{"3"} "Hockey stick" catch based on stock status
  }}

\item{hcr.opts}{\itemize{
\item \code{"percentFXSPR"} percent of F_XSPR to use for calculating catch in projections, default = 75
\item \code{"percentSPR"} X\% SPR used to calculate reference points, default = 40
\item \code{"avg.yrs"} controls which years the MAA, WAA, M, F will be averaged over in the projections, default = 5
\item \code{"max_percent"} maximum percent of F_XSPR to use for calculating catch in projections, default = 75
\item \code{"min_percent"} minimum percent of F_XSPR to use for calculating catch in projections, default = 0.01
\item \code{"BThresh_up"} Upper bound of overfished level, default = 0.5
\item \code{"BThresh_low"} Lower bound of overfished level, default = 0.1
}}
}
\value{
a list of catch advice
}
\description{
A function to generate catch advice for management strategy evaluation.
}
\examples{
\dontrun{
data <- generate_basic_info(n_stocks = 2, n_regions = 2, n_indices = 2, n_fleets = 2, base.years = 2003:2022)
NAA_re <- list(N1_model=c("equilibrium","equilibrium"),sigma=c("rec","rec"),cor=c("iid","iid"),recruit_model = 2)
input <- prepare_wham_input(basic_info = data, NAA_re = NAA_re)
mod <- fit_wham(input, do.fit = FALSE)
input$data = mod$simulate(complete=TRUE)
mod <- fit_wham(input, do.osa = FALSE, do.retro = FALSE, MakeADFun.silent = FALSE)
catch <- advice_fn(mod, pro.yr = 3, type = 1)
}
}
\seealso{
\code{\link{project_wham}}
}
