% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_fn.R
\name{advice_fn}
\alias{advice_fn}
\title{Generate Catch Advice for Management Strategy Evaluation}
\usage{
advice_fn(em, pro.yr = assess.interval, hcr = NULL)
}
\arguments{
\item{em}{A fitted WHAM estimation model.}

\item{pro.yr}{Integer. Number of years for projection.}

\item{hcr}{List. Harvest control rule:
\describe{
  \item{\code{hcr.type}}{Integer (1: F\_XSPR, 2: constant catch, 3: hockey stick).}
  \item{\code{hcr.opts}}{List of options, including:
    \itemize{
      \item \code{use_FXSPR}: Logical. Use F\%SPR-based projections (default: TRUE).
      \item \code{percentFXSPR}: Numeric. % of F\%SPR to apply (default: 75).
      \item \code{use_FMSY}: Logical. Use FMSY-based projections (default: FALSE).
      \item \code{percentFMSY}: Numeric. % of FMSY to apply (default: 75).
      \item \code{avg_yrs}: Integer. Number of years to average over (default: 5).
      \item \code{cont.M.re}: Logical. Continue M random effects (default: FALSE).
      \item \code{cont.move.re}: Logical. Continue movement random effects (default: NULL if 1 region).
      \item \code{max_percent}: Numeric. Max % of F\%SPR or FMSY (default: 75).
      \item \code{min_percent}: Numeric. Min % of F\%SPR or FMSY (default: 0.01).
      \item \code{BThresh_up}: Numeric. Upper biomass threshold (default: 0.5).
      \item \code{BThresh_low}: Numeric. Lower biomass threshold (default: 0.1).
    }
  }
}}
}
\value{
A matrix containing the projected catch advice for \code{pro.yr} years.
}
\description{
A function to generate catch advice based on different harvest control rules (HCRs)
for use in management strategy evaluation (MSE).
}
\examples{
# Example: Define a harvest control rule using F_40\%SPR with 100\% target
hcr <- list(
  hcr.type = 1,
  hcr.opts = list(
    use_FXSPR = TRUE,
    percentFXSPR = 100,  # Use 100\% of F_40\%SPR
    avg_yrs = 5,         # Average over the last 5 years
    cont.M.re = FALSE,   # Do not continue M random effects
    cont.move.re = FALSE # Do not continue movement random effects
  )
)

# Run advice_fn with a WHAM model `em`
# advice <- advice_fn(em, pro.yr = 5, hcr = hcr)

}
\seealso{
\code{\link{project_wham}}
}
