% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advice_fn.R
\name{advice_fn}
\alias{advice_fn}
\title{Generate Catch Advice for Management Strategy Evaluation}
\usage{
advice_fn(em, pro.yr = assess.interval, hcr)
}
\arguments{
\item{em}{An estimation model fitted using WHAM.}

\item{pro.yr}{Integer. The number of years for projection.}

\item{hcr}{A list specifying the harvest control rule (HCR) type and options. It should contain:
\itemize{
  \item{\code{hcr.type}}{Integer. The type of harvest control rule to use. Options include:}
  \itemize{
    \item \code{1}: Annual projected catch based on 75\% of F40\% (default).
    \item \code{2}: Constant catch based on 75\% of F40\%.
    \item \code{3}: "Hockey stick" approach based on stock status.
  }
  \item{\code{hcr.opts}}{A list of additional options for the harvest control rule, including:}
  \itemize{
    \item{\code{use.FXSPR}}{Logical. Calculate and use F at X\% SPR for projections (default = TRUE).}
    \item{\code{percentFXSPR}}{Numeric. The percentage of F\_XSPR to use for calculating catch in projections (default = 75).}
    \item{\code{use.FMSY}}{Logical. Calculate and use FMSY for projections (default = FALSE).}
    \item{\code{percentFMSY}}{Numeric. Percentage of F_MSY to use for catch projections. Used only if use.FMSY = TRUE (default = 100).}
    \item{\code{avg.yrs}}{Integer. The number of years over which mean values of MAA, WAA, M, and F will be averaged in projections (default = 5).}
    \item{\code{max_percent}}{Numeric. Maximum percentage of F\_XSPR to use in projections (default = 75).}
    \item{\code{min_percent}}{Numeric. Minimum percentage of F\_XSPR to use in projections (default = 0.01).}
    \item{\code{BThresh_up}}{Numeric. Upper bound of overfished biomass threshold (default = 0.5).}
    \item{\code{BThresh_low}}{Numeric. Lower bound of overfished biomass threshold (default = 0.1).}
    \item{\code{cont.M.re}}{Logical. Whether to continue natural mortality random effects (e.g., AR1_y or 2D AR1) during projections (default = FALSE).}
    \item{\code{cont.move.re}}{Logical. Whether to continue movement random effects during projections (default = FALSE).}
  }
}}
}
\value{
A matrix containing the projected catch advice for the specified number of projection years.
}
\description{
A function to generate catch advice based on different harvest control rules (HCRs) for use in management strategy evaluation.
}
\seealso{
\code{\link{project_wham}}
}
