% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_catch_advice.R
\name{calculate_catch_advice}
\alias{calculate_catch_advice}
\title{Calculate Catch Advice for Region- and Gear-Specific Fleets}
\usage{
calculate_catch_advice(
  om,
  advice,
  aggregate_catch_info,
  aggregate_index_info,
  final_year,
  catch_alloc = list(weight_type = 1, method = "equal", user_weights = NULL, weight_years
    = 1, survey_pointer = 1)
)
}
\arguments{
\item{om}{List. The operating model containing input data, fleet, and index information.}

\item{advice}{Matrix or vector. Catch advice of dimension (n_years, n_fleets), 
where each column represents a gear type. If advice is a vector, it represents total annual catch.}

\item{aggregate_catch_info}{List. Contains fleet allocation information:
\itemize{
  \item $fleet_pointer - Vector indicating the gear type assignment for each fleet.
}}

\item{aggregate_index_info}{List. Contains index region allocation information:
\itemize{
  \item $index_pointer - Vector indicating the index region assignment.
}}

\item{final_year}{Integer. The last year to consider when calculating historical catch weights.}

\item{catch_alloc}{List. Contains specifications for catch allocation:
\itemize{
  \item $weight_type - Integer (1–4) indicating the type of weighting used.
    \itemize{
      \item 1 - Equal weighting across fleets.
      \item 2 - Weighting based on past catch data.
      \item 3 - Weighting based on past survey index data.
      \item 4 - User-specified weights.
    }
  \item $method - String. Specifies the catch allocation method:
    \itemize{
      \item "equal" - Total catch is equally allocated to all fleets across regions.
      \item "gear_equal" - Gear-specific total catch is equally allocated to fleets using the same gear type.
      \item "fleet_region" - Total catch is allocated to each region based on historical regional catch. Then, the regional catch is equally distributed among all fleets within the region.
      \item "fleet_gear" - Total catch is allocated to each gear type based on historical gear-specific catch. Then, the gear-specific catch is equally distributed among regions that use this gear.
      \item "fleet_combined" - Total catch is allocated based on historical regional and gear-specific catch.
      \item "fleet_catch" - Total catch is allocated based on historical fleet-specific catch within each region.
      \item "index_equal" - Total catch is allocated to each region based on historical survey catch. Then, the regional catch is distributed among fleets based on historical gear-specific catch.
      \item "index_gear" - Total catch is allocated based on historical region-specific survey catch and then distributed among fleets based on historical gear-specific catch.
      \item "multiple_index_equal" - Total catch is allocated to each region based on historical survey catch from multiple surveys. Then, the regional catch is equally distributed among fleets within the region.
      \item "multiple_index_gear" - Total catch is allocated to each region based on multiple historical survey-based catch estimates, then distributed among fleets within the region based on historical gear-specific catch.
      \item "user_defined_fleets" - Catch allocation is based on user-defined weights for each fleet.
      \item "user_defined_regions" - Catch is allocated based on user-defined weights for regions, then equally distributed among fleets within the region.
    }
  \item $user_weights - Numeric vector (n_regions or n_fleets). Optional. User-defined weights summing to 1 (use when weight_type = 4).
  \item $weight_years - Integer. Number of years to average for calculating historical catch weights.
  \item $survey_pointer - Integer/Vector. Specifies which survey index type to use for weighting (use when weight_type = 3).
}}
}
\value{
A matrix of dimension (n_years, n_fleets) representing region-specific 
  and gear-specific catch advice.
}
\description{
This function distributes total catch into region- and gear-specific fleets 
based on different weighting methods: "fleet_region", "fleet_gear", "fleet_combined", "fleet_catch", "index_equal", "index_gear", "multiple_index_equal", "multiple_index_gear", or user-specified weights.
}
