% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_info.R
\name{generate_basic_info}
\alias{generate_basic_info}
\title{Generate basic information for multi-wham input}
\usage{
generate_basic_info(
  n_stocks = 2,
  n_regions = 2,
  n_indices = 2,
  n_fleets = 2,
  n_seasons = 4,
  base.years = 1993:2022,
  n_feedback_years = 0,
  life_history = "medium",
  n_ages = 12,
  Fbar_ages = 12,
  recruit_model = 2,
  F_info = list(F.year1 = 0.2, Fhist = "constant", Fmax = 0.2, Fmin = 0.2, change_time =
    0.5),
  catch_info = list(catch_cv = 0.1, catch_Neff = 100),
  index_info = list(index_cv = 0.1, index_Neff = 100, fracyr_indices = 0.625, q = 0.2),
  fracyr_spawn = 0.625,
  bias.correct.process = FALSE,
  bias.correct.observation = FALSE,
  bias.correct.BRPs = FALSE,
  mig_type = 0,
  XSPR_R_opt = 2
)
}
\arguments{
\item{n_stocks}{Number of stocks}

\item{n_regions}{Number of regions}

\item{n_indices}{Number of indices}

\item{n_fleets}{Number of fleets}

\item{n_seasons}{Number of seasons}

\item{base.years}{Model years (or "burn-in" period in MSE)}

\item{n_feedback_years}{(optional) Number of years in the MSE feedback loop}

\item{life_history}{Fish life history (parameters are obtained from \href{https://doi.org/10.1139/cjfas-2016-0381}{Wiedenmann et al. 2017}):
\itemize{
  \item \code{"median"} median-lived species (default)
  \item \code{"short"} short-lived species
  \item \code{"long"} long-lived species
  }}

\item{n_ages}{Number of ages}

\item{Fbar_ages}{Ages to use to average total F at age defining fully selected F for reference points.}

\item{recruit_model}{Option to specify stock-recruit model
\itemize{
  \item \code{1} {SCAA (without NAA_re option specified) or Random walk (if NAA_re$sigma specified), i.e. predicted recruitment in year i = recruitment in year i-1}
  \item \code{2} {(default) Random about mean, i.e. steepness = 1}
  \item \code{3} {Beverton-Holt}
  \item \code{4} {Ricker}
}}

\item{F_info}{Historical fishing pressure
\itemize{
  \item \code{$F.year1} fishing mortality in the first year
  \item \code{$Fhist}: \cr
  {"constant"} = constant across years \cr
  {"updown"}   = increase from F.year1 to Fmax until a change point and then decrease to Fmin \cr
  {"downup"}   = decrease from F.year1 to Fmin until a change point and then increase to Fmax \cr
  {"F-H-L"} = constant Fmax(a multiplier)xF.year1 until the change point and then constant Fmin(a multiplier)xF.year1
  \item \code{$Fmax} maximum F (or a multiplier when Fhist = "F-H-L")
  \item \code{$Fmin} minimum F (or a multiplier when Fhist = "F-H-L")
  \item \code{$change_time} a ratio that determines the change point
  }}

\item{catch_info}{Fleet information
\itemize{
  \item \code{$catch_cv} cv for the catch
  \item \code{$catch_Neff} effective sample size for the catch
  }}

\item{index_info}{Survey information (see details)
\itemize{
  \item \code{$index_cv} cv for the indices
  \item \code{$index_Neff} effective sample size for the indices
  \item \code{$fracyr_indices} fraction of the year when survey is conducted
  }}

\item{fracyr_spawn}{Fraction of the year when spawning occurs}

\item{bias.correct.process}{T/F process error is bias corrected}

\item{bias.correct.observation}{T/F observation error is bias corrected}

\item{bias.correct.BRPs}{T/F biological reference point is bias corrected}

\item{mig_type}{0 = migration after survival, 1 = movement and mortality simultaneous}

\item{XSPR_R_opt}{\itemize{
\item \code{1} use annual R estimates for annual SSB_XSPR
\item \code{2} use average R estimates for annual SSB_XSPR (default)
\item \code{3} use annual expected R for annual SSB_XSPR
\item \code{4} use average expected R estimates for annual SSB_XSPR
\item \code{5} use bias-corrected expected recruitment
}}

\item{q}{Survey catchability}
}
\value{
A list of information that will be used to prepare wham input
}
\description{
The generate_basic_info() function is used to generate biological and fishery information for 
simulation-estimation and management strategy evaluation. 
Note that not all information is included here. More details can be found in \code{\link{prepare_wham_input}}
}
\examples{
\dontrun{
data <- generate_basic_info()
}
}
\seealso{
\code{\link{prepare_wham_input}}
}
