% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ini_N1.R
\name{generate_ini_N1}
\alias{generate_ini_N1}
\title{Generate initial numbers-at-age}
\usage{
generate_ini_N1(
  basic_info,
  ini.opt = "equilibrium",
  log_N1 = NULL,
  log_N1_F = NULL,
  init_NAA = NULL
)
}
\arguments{
\item{basic_info}{list specifying options for numbers-at-age random effects, initial parameter values, and recruitment model (see details)}

\item{ini.opt}{N1_model}

\item{log_N1}{a vector (length = n_stocks) of log number-at-age 1 at equilibrium}

\item{log_N1_F}{a vector (length = n_stocks) of fishing mortality at equilibrium
\describe{
   \item{"age-specific-fe"}{(default) age- and region-specific fixed effects parameters}
   \item{"equilibrium"}{2 fixed effects parameters: an initial recruitment and an instantaneous fishing mortality rate to generate an equilibruim abundance at age.}
   \item{"iid-re"}{(default) age- and region-specific iid random effects parameters. 2 parameters: mean and sd for log NAA}
   \item{"ar1-re"}{(default) age- and region-specific random effects parameters. 3 parameters: mean and sd, and cor for log NAA}
}}

\item{init_NAA}{a matrix (n_ages x n_stocks) of initial numbers-at-age}
}
\value{
an array of numbers-at-age (n_stocks x n_regions x n_ages)
}
\description{
Generate initial numbers-at-age
}
