% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_move.R
\name{generate_move}
\alias{generate_move}
\title{Configure movement random effects}
\usage{
generate_move(
  basic_info,
  move.type = 2,
  move.rate = c(0.3, 0.1),
  move.re = "constant",
  move.sigma = 0.2,
  prior.sigma = 0.2,
  move.rho_a = 0.5,
  move.rho_y = 0.5,
  use.prior = FALSE,
  move_dyn = 0
)
}
\arguments{
\item{basic_info}{Information generated by \code{\link{generate_basic_info}}.
Note: NAA_where has to be specified in basic_info using \code{\link{generate_NAA_where}}.}

\item{move.type}{Movement type
\itemize{
  \item \code{1} single stock movement (i.e. only stock 1 can move), with region-specific movement rates
  \item \code{2} bidirectional movement (i.e. all stocks can move), with region-specific movement rates (default)
  \item \code{3} no movement
}}

\item{move.rate}{Movement rates (default = c(0.3, 0.1))}

\item{move.re}{Movement random effects
\itemize{
  \item \code{"constant"} constant movement rate across years and ages (default).
  \item \code{"iid_y"} movement rate deviations vary by year, but uncorrelated (IID).
  \item \code{"iid_a"} movement rate deviations vary by age, but uncorrelated (IID).
  \item \code{"ar1_y"} movement rate deviations correlated by year (AR1).
  \item \code{"ar1_a"} movement rate deviations correlated by age (AR1).
  \item \code{"iid_y","iid_a"} vector, movement rate deviations vary by year and ages, but uncorrelated (2D-IID).
  \item \code{"ar1_y","ar1_a"} vector, movement rate deviations vary by year and ages (2D-AR1).
}}

\item{move.sigma}{Standard deviation for movement random effects (default = 0.2)}

\item{prior.sigma}{Standard deviation for normal priors on mean movement (default = 0.2)}

\item{move.rho_a}{Correlation for movement random effects by ages, only used if move.re = "ar1_a" (default = 0.5)}

\item{move.rho_y}{Correlation for movement random effects by years, only used if move.re = "ar1_y" (default = 0.5)}

\item{use.prior}{Logical, whether to use priors for movement estimation (default = FALSE)}

\item{move_dyn}{Movement dynamics. 0 = natal homing, 1 = meta-population}
}
\value{
A named list with the following components:
  \describe{
    \item{$stock_move}{Length = n_stocks, TRUE/FALSE whether each stock can move. If not provided, movement will be defined below for all stocks.}
    \item{$separable}{Length = n_stocks, TRUE/FALSE whether movement should be modeled separably from mortality or both occurring simultaneously.}
    \item{$must_move}{Array (n_stocks x n_seasons x n_regions) 0/1 indicator whether movement from a region must occur.}
    \item{$can_move}{Array (n_stocks x n_seasons x n_regions x n_regions) 0/1 indicator whether movement can occur from one region to another.}
    \item{$mean_vals}{Array (n_stocks x n_seasons x n_regions x n_regions - 1) of initial movement rate parameters from each region.}
    \item{$sigma_vals}{Array (n_stocks x n_seasons x n_regions x n_regions - 1) of initial standard deviations for random effects.}
    \item{$cor_vals}{Array (n_stocks x n_seasons x n_regions x n_regions - 1 x 2) of initial correlation values for random effects.}
    \item{$use_prior}{Array (n_stocks x n_seasons x n_regions x n_regions - 1) 0/1 indicator whether to include prior for mean movement parameters in joint log-likelihood.}
    \item{$prior_sigma}{Array (n_stocks x n_seasons x n_regions x n_regions - 1) of standard deviation parameters for normal priors on mean movement parameters on transformed scale (-Inf,Inf).}
  }
}
\description{
The `generate_move()` function is used to configure movement random effects.
}
\examples{
\dontrun{
data <- generate_basic_info()
data <- generate_move(data, move.type = 2, move.rate = c(0.3,0.2), move.re = "constant")
}
}
\seealso{
\code{\link{set_move}} and \code{\link{generate_NAA_where}}
}
