% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_through_fn.R
\name{loop_through_fn}
\alias{loop_through_fn}
\title{Perform Management Strategy Evaluation (MSE)}
\usage{
loop_through_fn(
  om,
  em_info = NULL,
  random = NULL,
  M_em = NULL,
  sel_em = NULL,
  NAA_re_em = NULL,
  move_em = NULL,
  age_comp_em = NULL,
  em.opt = list(separate.em = TRUE, separate.em.type = 1, do.move = FALSE, est.move =
    FALSE),
  global_waa = TRUE,
  aggregate_catch_info = NULL,
  aggregate_index_info = NULL,
  filter_indices = NULL,
  assess_years = NULL,
  assess_interval = NULL,
  base_years = NULL,
  year.use = 20,
  hcr = list(hcr.type = 1, hcr.opts = NULL),
  catch_alloc = list(weight_type = 1, method = "region", user_weights = NULL,
    weight_years = 1),
  do.retro = FALSE,
  do.osa = FALSE,
  seed = 123,
  save.sdrep = FALSE,
  save.last.em = FALSE
)
}
\arguments{
\item{om}{A fitted operating model containing pseudo data.}

\item{em_info}{A list containing information used to generate the estimation model (default = NULL).}

\item{random}{Character vector of processes treated as random effects in the operating model (default = "log_NAA").}

\item{M_em}{Configuration of natural mortality in the estimation model.}

\item{sel_em}{Configuration of selectivity in the estimation model.}

\item{NAA_re_em}{Configuration of numbers-at-age (NAA) in the estimation model.}

\item{move_em}{Configuration of movement in the estimation model.}

\item{age_comp_em}{Character. The likelihood distribution of age composition data in the estimation model. Options include:
\itemize{
  \item \code{"multinomial"} (default)
  \item \code{"dir-mult"}
  \item \code{"dirichlet-miss0"}
  \item \code{"dirichlet-pool0"}
  \item \code{"logistic-normal-miss0"}
  \item \code{"logistic-normal-ar1-miss0"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl"}
  \item \code{"logistic-normal-01-infl-2par"}
  \item \code{"mvtweedie"}
  \item \code{"dir-mult-linear"}
}}

\item{em.opt}{List of options for the assessment model, including:
\describe{
  \item{\code{separate.em}}{Logical. If TRUE, separate estimation models are used for different stocks (default = FALSE).}
  \item{\code{separate.em.type}}{Integer. Used only if \code{separate.em = TRUE}:}
    \itemize{
      \item 1 = Panmictic (spatially-aggregated model).
      \item 2 = Fleets-as-areas model.
      \item 3 = Separate assessment models for each region (n = number of regions).
    }
  \item{\code{do.move}}{Logical. If TRUE, movement is included (default = FALSE).}
  \item{\code{est.move}}{Logical. If TRUE, movement rates are estimated (default = FALSE).}
}}

\item{aggregate_catch_info}{List. Contains details for aggregating fleet-specific catch data:
\itemize{
  \item `fleet_pointer` - Integer vector of length `n_fleets`, specifying how fleets should be aggregated.
  \item `catch_cv`, `catch_Neff` - Numeric vectors of length `n_fleets`, specifying catch CV and sample size.
  \item `use_agg_catch`, `use_catch_paa` - Boolean vectors of length `n_fleets`, indicating whether to aggregate.
}}

\item{aggregate_index_info}{List. Contains details for aggregating index-specific data:
\itemize{
  \item `index_pointer` - Integer vector of length `n_indices`, specifying how indices should be aggregated.
  \item `index_cv`, `index_Neff`, `q` - Numeric vectors of length `n_indices`, specifying index CV, effective sample size, and catchability coefficient.
  \item `use_indices`, `use_index_paa` - Boolean vectors of length `n_indices`, indicating whether to aggregate.
}}

\item{filter_indices}{Integer (0/1) vector (optional). User-specified which indices are excluded from the assessment model. For example, c(1,0,1,1) indicates Index 1 (include) and 2 (exclude) in region 1, Index 3 (include) and 4 (include) in region 2}

\item{assess_years}{Vector of years when assessments are conducted.}

\item{assess_interval}{Integer. The interval between stock assessments in the MSE feedback loop.}

\item{base_years}{Vector of years used in the burn-in period.}

\item{year.use}{Integer. Number of years included in the estimation model (default = 30).}

\item{hcr}{List containing harvest control rule (HCR) settings:
\describe{
  \item{\code{hcr.type}}{Integer. The type of harvest control rule:}
    \itemize{
      \item \code{1} Annual projected catch based on 75\% of F40\% (default).
      \item \code{2} Constant catch based on 75\% of F40\%.
      \item \code{3} "Hockey stick" catch based on stock status.
    }
  \item{\code{hcr.opts}}{(Only for \code{hcr.type = 3}) A list containing additional HCR options:}
    \describe{
      \item{\code{max_percent}}{Maximum percentage of F\_XSPR used for catch projections (default = 75).}
      \item{\code{min_percent}}{Minimum percentage of F\_XSPR used for catch projections (default = 0.01).}
      \item{\code{BThresh_up}}{Upper bound of overfished biomass threshold (default = 0.5).}
      \item{\code{BThresh_low}}{Lower bound of overfished biomass threshold (default = 0.1).}
    }
}}

\item{catch_alloc}{List. Contains specifications for catch allocation:
\itemize{
  \item `$weight_type` - Integer (1–4) indicating the type of weighting used.
    \itemize{
      \item `1` - Equal weighting across fleets.
      \item `2` - Weighting based on past catch data.
      \item `3` - Weighting based on past survey index data.
      \item `4` - User-specified weights.
    }
  \item `$method` - String. Specifies the catch allocation method:
    \itemize{
      \item `"region"` - Uses regional catch totals to compute weights.
      \item `"gear"` - Uses gear-type catch totals to compute weights.
      \item `"combined"` - First allocates to gear-specific total catch, then splits into regions.
    }
  \item `$user_weights` - Numeric vector (`n_regions`). Optional. User-defined weights summing to 1.
  \item `$weight_years` - Integer. Number of years to average for calculating historical catch weights.
}}

\item{do.retro}{Logical. If TRUE, performs retrospective analysis (default = TRUE).}

\item{do.osa}{Logical. If TRUE, calculates one-step-ahead (OSA) residuals (default = TRUE).}

\item{seed}{Integer. The random seed used for stochastic processes.}

\item{save.sdrep}{Logical. If TRUE, saves the results of every assessment model (memory-intensive).}

\item{save.last.em}{Logical. If TRUE, saves only the last estimation model (default = FALSE).}

\item{ind_em}{Vector. Indices specifying the years for which the estimation model should use data.}
}
\value{
A list containing:
  \describe{
    \item{\code{om}}{Updated operating model.}
    \item{\code{em_list}}{List of estimation model results across assessment years.}
    \item{\code{par.est}}{Estimated parameter values.}
    \item{\code{par.se}}{Standard errors of parameter estimates.}
    \item{\code{adrep.est}}{Estimated values from ADMB report.}
    \item{\code{adrep.se}}{Standard errors from ADMB report.}
    \item{\code{opt_list}}{Optimization results from WHAM.}
    \item{\code{converge_list}}{Convergence diagnostics.}
    \item{\code{catch_advice}}{Projected catch advice over assessment years.}
    \item{\code{em_full}}{List of full estimation model outputs.}
    \item{\code{runtime}}{Elapsed time for function execution.}
  }
}
\description{
A wrapper function to iterate through the feedback period in a management strategy evaluation (MSE), 
updating the operating model (OM), refitting the estimation model (EM), and generating catch advice.
}
\seealso{
\code{\link{make_em_input}}, \code{\link{update_om_fn}}, \code{\link{advice_fn}}
}
