% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_through_fn.R
\name{loop_through_fn}
\alias{loop_through_fn}
\title{Perform Management Strategy Evaluation (MSE)}
\usage{
loop_through_fn(
  om,
  em_info = NULL,
  random = NULL,
  M_em = NULL,
  sel_em = NULL,
  NAA_re_em = NULL,
  move_em = NULL,
  age_comp_em = "multinomial",
  em.opt = list(separate.em = TRUE, separate.em.type = 1, do.move = FALSE, est.move =
    FALSE),
  aggregate_catch_info = NULL,
  aggregate_index_info = NULL,
  aggregate_weights_info = NULL,
  reduce_region_info = NULL,
  filter_indices = NULL,
  update_catch_info = NULL,
  update_index_info = NULL,
  assess_years = NULL,
  assess_interval = NULL,
  base_years = NULL,
  year.use = 20,
  add.years = FALSE,
  by_fleet = TRUE,
  FXSPR_init = NULL,
  hcr = list(hcr.type = 1, hcr.opts = NULL),
  catch_alloc = list(weight_type = 1, method = "equal", user_weights = NULL, weight_years
    = 1),
  do.retro = FALSE,
  do.osa = FALSE,
  do.brps = FALSE,
  seed = 123,
  save.sdrep = FALSE,
  save.last.em = FALSE
)
}
\arguments{
\item{om}{An operating model that contains pseudo data.}

\item{em_info}{A list that contains information used to generate the assessment model (default = NULL).}

\item{random}{Character vector of processes treated as random effects in the operating model (default = "log_NAA").}

\item{M_em}{Configuration of natural mortality in the assessment model.}

\item{sel_em}{Configuration of selectivity in the assessment model.}

\item{NAA_re_em}{Configuration of numbers-at-age (NAA) in the assessment model.}

\item{move_em}{Configuration of movement in the assessment model.}

\item{age_comp_em}{Character. The likelihood distribution of age composition data in the assessment model. Options include:
\itemize{
  \item \code{"multinomial"} (default)
  \item \code{"dir-mult"}
  \item \code{"dirichlet-miss0"}
  \item \code{"dirichlet-pool0"}
  \item \code{"logistic-normal-miss0"}
  \item \code{"logistic-normal-ar1-miss0"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl"}
  \item \code{"logistic-normal-01-infl-2par"}
  \item \code{"mvtweedie"}
  \item \code{"dir-mult-linear"}
}}

\item{em.opt}{List of options for the assessment model, including:
\describe{
  \item{\code{separate.em}}{Logical. Whether to separate assessment models (default = FALSE).}
  \item{\code{separate.em.type}}{Integer. Used only if \code{separate.em = TRUE}:}
    \itemize{
      \item 1 = Panmictic (spatially-aggregated model).
      \item 2 = Spatially implicit assessment model (fleets-as-areas).
      \item 3 = Separate region-specific assessment models (n = number of regions).
    }
  \item{\code{do.move}}{Logical. Whether to include movement in the assessment model (default = FALSE).}
  \item{\code{est.move}}{Logical. Whether to estimate movement rates in the assessment model (default = FALSE).}
}}

\item{aggregate_catch_info}{List (optional). User-specified catch aggregation settings for panmictic models using aggregate catch.
\itemize{
  \item `$n_fleets` Integer. Number of fleets.
  \item `$catch_cv` Numeric vector (`n_fleets`). CVs for annual aggregate catches by fleet.
  \item `$catch_Neff` Numeric vector (`n_fleets`). Effective sample sizes for fleet catches.
  \item `$use_agg_catch` Integer vector (`n_fleets`). 0/1 values flagging whether to use aggregate catches.
  \item `$use_catch_paa` Integer vector (`n_fleets`). 0/1 values flagging whether to use proportions at age observations.
  \item `$fleet_pointer` Integer vector (`n_fleets`). Defines fleet grouping (0 = exclude).
  \item `$use_catch_weighted_waa` Logical. Whether to use weighted weight-at-age based on fleet catches.
}}

\item{aggregate_index_info}{List (optional). User-specified index aggregation settings for panmictic models using aggregate indices.
\itemize{
  \item `$n_indices` Integer. Number of indices.
  \item `$index_cv` Numeric vector (`n_indices`). CVs for annual aggregate index catches.
  \item `$index_Neff` Numeric vector (`n_indices`). Effective sample sizes for survey indices.
  \item `$fracyr_indices` Numeric vector (`n_indices`). Fraction of the year for each survey index.
  \item `$q` Numeric vector (`n_indices`). Initial survey catchabilities.
  \item `$use_indices` Integer vector (`n_indices`). 0/1 values flagging whether to use survey indices.
  \item `$use_index_paa` Integer vector (`n_indices`). 0/1 values flagging whether to use proportions at age observations.
  \item `$units_indices` Integer vector (`n_indices`). 1/2 values flagging whether aggregate observations are biomass (1) or numbers (2).
  \item `$units_index_paa` Integer vector (`n_indices`). 1/2 values flagging whether composition observations are biomass (1) or numbers (2).
  \item `$index_pointer` Integer vector (`n_indices`). Defines index grouping (0 = exclude).
  \item `$use_index_weighted_waa` Logical. Whether to use weighted weight-at-age based on survey catches.
   }}

\item{aggregate_weights_info}{List (optional). Specifies how to compute weighted averages for
weight-at-age (`waa`) and maturity-at-age during data aggregation (For panmictic and fleets-as-areas models only).
Used for aggregating across fleets or indices to form total/stock/regional summaries.
\itemize{
  \item `$ssb_waa_weights` List. Settings for weighting weight-at-age used for spawning stock biomass.
    \itemize{
      \item `$fleet` Logical. Whether to use fleet-specific weights.
      \item `$index` Logical. Whether to use index-specific weights.
      \item `$pointer` Integer. Index pointing to the selected fleet or index group (e.g., 1 means the first valid fleet group).
    }
  \item `$maturity_weights` List. Settings for weighting maturity-at-age used for spawning stock biomass.
    \itemize{
      \item `$fleet` Logical. Whether to use fleet-specific weights.
      \item `$index` Logical. Whether to use index-specific weights.
      \item `$pointer` Integer. Index pointing to the selected fleet or index group.
    }
}}

\item{reduce_region_info}{List (optional). Remove specific regions from the assessment model. If `NULL`, no modifications are applied.
The expected components include:
\itemize{
  \item `$remove_regions` Integer vector (`n_regions`). 0/1 values flagging whether to include regions.
  \item `$reassign` Numeric. Specify reassignment for surveys from the "removed" regions to "remaining" regions.
  \item `$NAA_where`Integer array (`n_stocks × n_regions × n_ages`) 0/1 values flagging whether the stock from a certain age can be present in a specific region.
  \item `$sel_em` Selectivity configuration for the "reduced" assessment model.
  \item `$M_em` Natural mortality configuration for the "reduced" assessment model.
  \item `$NAA_re_em` Numbers-at-age configuration for the "reduced" assessment model.
  \item `$move_em` Movement configuration for the "reduced" assessment model.
  \item `$onto_move_list` List of ontogenetic movement information for the "reduced" assessment model:
    \itemize{
      \item `$onto_move` (array, dimension: `n_stocks × n_regions × (n_regions-1)`)  
        Age-specific movement type.
      \item `$onto_move_pars` (array, dimension: `n_stocks × n_regions × (n_regions-1) × 4`)  
        Parameters controlling age-specific movement patterns.
      \item `$age_mu_devs` (array, dimension: `n_stocks × n_regions × (n_regions-1) × n_ages`)  
        User-specified deviations in mean movement rates across ages.
    }
}}

\item{filter_indices}{Integer (0/1) vector (optional). User-specified which indices are excluded from the assessment model. For example, c(1,0,1,1) indicates Index 1 (include) and 2 (exclude) in region 1, Index 3 (include) and 4 (include) in region 2}

\item{update_catch_info}{List (optional). Update catch CV and Neff in the EM.
The expected components include:
\itemize{
  \item `$agg_catch_sigma` Matrix. Either full (n_years x n_fleets) or subset (length(ind_em) x n_fleets).
  \item `$catch_Neff` Matrix. Same dimension as `agg_catch_sigma`.
}}

\item{update_index_info}{List (optional). Update index CV and Neff in the EM.
The expected components include:
\itemize{
  \item `$agg_index_sigma` Matrix. Either full (n_years x n_indices) or subset (length(ind_em) x n_indices).
  \item `$index_Neff` Matrix. Same dimension as `agg_index_sigma`.
}}

\item{assess_years}{Vector of years when assessments are conducted.}

\item{assess_interval}{Integer. The interval between assessments in the MSE feedback loop.}

\item{base_years}{Vector of years used in the burn-in period.}

\item{year.use}{Integer. Number of years included in the assessment model (default = 20).}

\item{add.years}{Logical. Whether to use entire time series of data in the assessment model (default = FALSE).}

\item{by_fleet}{Logical. Whether to calculate fleet-specific fishing mortality (default = TRUE).}

\item{FXSPR_init}{Numeric. Change initial fishing mortality for estimating reference point in the assessment model (default = NULL).}

\item{hcr}{A list specifying the harvest control rule (HCR) type and options. It should contain:
\itemize{
  \item{\code{hcr.type}}{Integer. The type of harvest control rule to use. Options include:}
  \itemize{
    \item \code{1}: Annual projected catch based on 75\% of F40\% (default).
    \item \code{2}: Constant catch based on 75\% of F40\%.
    \item \code{3}: "Hockey stick" approach based on stock status.
  }
  \item{\code{hcr.opts}}{A list of additional options for the harvest control rule, including:}
  \itemize{
    \item{\code{use.FXSPR}}{Logical. Calculate and use F at X\% SPR for projections (default = TRUE).}
    \item{\code{percentSPR}}{Numeric. The X\% SPR used to calculate reference points (default = 40).}
    \item{\code{percentFXSPR}}{Numeric. The percentage of F\_XSPR to use for calculating catch in projections (default = 75).}
    \item{\code{use.FMSY}}{Logical. Calculate and use FMSY for projections (default = FALSE).}
    \item{\code{percentFMSY}}{Numeric. Percentage of F_MSY to use for catch projections. Used only if use.FMSY = TRUE (default = 100).}
    \item{\code{avg.yrs}}{Integer. The number of years over which mean values of MAA, WAA, M, and F will be averaged in projections (default = 5).}
    \item{\code{max_percent}}{Numeric. Maximum percentage of F\_XSPR to use in projections (default = 75).}
    \item{\code{min_percent}}{Numeric. Minimum percentage of F\_XSPR to use in projections (default = 0.01).}
    \item{\code{BThresh_up}}{Numeric. Upper bound of overfished biomass threshold (default = 0.5).}
    \item{\code{BThresh_low}}{Numeric. Lower bound of overfished biomass threshold (default = 0.1).}
    \item{\code{cont.M.re}}{Logical. Whether to continue natural mortality random effects (e.g., AR1_y or 2D AR1) during projections (default = FALSE).}
    \item{\code{cont.move.re}}{Logical. Whether to continue movement random effects during projections (default = FALSE).}
  }
}}

\item{catch_alloc}{List that contains specifications for catch allocation:
\itemize{
  \item `$weight_type` - Integer (1–4) indicating the type of weighting used.
    \itemize{
      \item `1` - Equal weighting.
      \item `2` - Weighting based on past catch data.
      \item `3` - Weighting based on past survey index data.
      \item `4` - User-specified weights.
    }
  \item `$method` - String. Specify the catch allocation method:
    \itemize{
      \item `"equal"` - Use equal weights, split total catch into fleet-specific catches (use when weight_type = 1).
      \item `"fleet_equal"` - Use equal weights, split total fleet catch into fleet-specific catches (use when weight_type = 1).
      \item `"fleet_region"` - Use regional catch totals to compute weights (use when weight_type = 2).
      \item `"fleet_gear"` - Use gear-type catch totals to compute weights (use when weight_type = 2).
      \item `"fleet_combined"` - First allocates to gear-specific total catch, then splits into regions (use when weight_type = 2).
      \item `"fleet_catch"` - Use fleet-specific catch to compute weights (use when weight_type = 2).
      \item `"index_equal"` - Use survey-based regional weighting, then assigns equally among fleets in the same region (use when weight_type = 3).
      \item `"index_gear"` - Use survey-based regional weighting, then assigns based on gear-specific weights (use when weight_type = 3).
      \item `"index_multiple"` - Use multiple survey-based regional weighting, then assigns equally among fleets in the same region (use when weight_type = 3).
      \item `"user_defined_fleets"` - Catch allocation is based on user-defined weights for each fleet.
      \item `"user_defined_regions"` - Catch is allocated based on user-defined weights for regions, then equally distributed among fleets within the region.
    }
  \item `$user_weights` - Numeric vector (`n_regions` or `n_fleets`). Optional. User-defined weights summing to 1 (use when weight_type = 4).
  \item `$weight_years` - Integer. Number of years to average for calculating historical catch weights.
  \item `$survey_pointer` - Integer vector. Specify which survey index type to use for weighting (use when weight_type = 3).
}}

\item{do.retro}{Logical. Whether to perform retrospective analysis for the assessment model (default = FALSE).}

\item{do.osa}{Logical. Whether to calculate one-step-ahead (OSA) residuals for the assessment model (default = FALSE).}

\item{do.brps}{Logical. Whether to calculate reference points in the operating model (default = FALSE).}

\item{seed}{Integer. The random seed used for stochastic processes.}

\item{save.sdrep}{Logical. Whether to save the results of every assessment model (memory-intensive).}

\item{save.last.em}{Logical. Whether to save the last estimation model (default = FALSE).}

\item{agg_index_sigma}{Matrix. Either full (n_years x n_indices) or subset (length(ind_em) x n_indices).}

\item{index_Neff}{Matrix. Same dimension as `agg_index_sigma`.}
}
\value{
A list containing:
  \describe{
    \item{\code{om}}{Updated operating model.}
    \item{\code{em_list}}{List of estimation model results across assessment years.}
    \item{\code{par.est}}{Estimated parameter values.}
    \item{\code{par.se}}{Standard errors of parameter estimates.}
    \item{\code{adrep.est}}{Estimated values from ADMB report.}
    \item{\code{adrep.se}}{Standard errors from ADMB report.}
    \item{\code{opt_list}}{Optimization results from WHAM.}
    \item{\code{converge_list}}{Convergence diagnostics.}
    \item{\code{catch_advice}}{Projected catch advice over assessment years.}
    \item{\code{em_full}}{List of full estimation model outputs.}
    \item{\code{runtime}}{Elapsed time for function execution.}
    \item{\code{seed.save}}{Random seed.}
    \item{\code{em_input}}{List of input information for the last assessment.}
  }
}
\description{
A wrapper function to iterate through the feedback period in a management strategy evaluation (MSE), 
updating the operating model (OM), refitting the assessment model (AM), and generating catch advice.
}
\seealso{
\code{\link{make_em_input}}, \code{\link{update_om_fn}}, \code{\link{advice_fn}}
}
