% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aggregate_data.R
\name{make_aggregate_data}
\alias{make_aggregate_data}
\title{Aggregate Data for Estimation Model}
\usage{
make_aggregate_data(
  om,
  em_info,
  ind_em,
  aggregate_catch_info,
  aggregate_index_info,
  aggregate_weights_info
)
}
\arguments{
\item{om}{List. The operating model containing observed data, including:}

\item{em_info}{List. The estimation model information, including:}

\item{ind_em}{Vector. Indices specifying the years for which the estimation model should use data.}

\item{aggregate_catch_info}{List (optional). User-specified catch aggregation settings for panmictic models using aggregate catch.
\itemize{
  \item `$n_fleets` Integer. Number of fleets.
  \item `$catch_cv` Numeric vector (`n_fleets`). CVs for annual aggregate catches by fleet.
  \item `$catch_Neff` Numeric vector (`n_fleets`). Effective sample sizes for fleet catches.
  \item `$use_agg_catch` Integer vector (`n_fleets`). 0/1 values flagging whether to use aggregate catches.
  \item `$use_catch_paa` Integer vector (`n_fleets`). 0/1 values flagging whether to use proportions at age observations.
  \item `$fleet_pointer` Integer vector (`n_fleets`). Defines fleet grouping (0 = exclude).
  \item `$use_catch_weighted_waa` Logical. Whether to use weighted weight-at-age based on fleet catches.
}}

\item{aggregate_index_info}{List (optional). User-specified index aggregation settings for panmictic models using aggregate indices.
\itemize{
  \item `$n_indices` Integer. Number of indices.
  \item `$index_cv` Numeric vector (`n_indices`). CVs for annual aggregate index catches.
  \item `$index_Neff` Numeric vector (`n_indices`). Effective sample sizes for survey indices.
  \item `$fracyr_indices` Numeric vector (`n_indices`). Fraction of the year for each survey index.
  \item `$q` Numeric vector (`n_indices`). Initial survey catchabilities.
  \item `$use_indices` Integer vector (`n_indices`). 0/1 values flagging whether to use survey indices.
  \item `$use_index_paa` Integer vector (`n_indices`). 0/1 values flagging whether to use proportions at age observations.
  \item `$units_indices` Integer vector (`n_indices`). 1/2 values flagging whether aggregate observations are biomass (1) or numbers (2).
  \item `$units_index_paa` Integer vector (`n_indices`). 1/2 values flagging whether composition observations are biomass (1) or numbers (2).
  \item `$index_pointer` Integer vector (`n_indices`). Defines index grouping (0 = exclude).
  \item `$use_index_weighted_waa` Logical. Whether to use weighted weight-at-age based on survey catches.
   }}

\item{aggregate_weights_info}{List (optional). Specifies how to compute weighted averages for
weight-at-age (`waa`) and maturity-at-age during data aggregation (For panmictic and fleets-as-areas models only).
Used for aggregating across fleets or indices to form total/stock/regional summaries.
\itemize{
  \item `$ssb_waa_weights` List. Settings for weighting weight-at-age used for spawning stock biomass.
    \itemize{
      \item `$fleet` Logical. Whether to use fleet-specific weights.
      \item `$index` Logical. Whether to use index-specific weights.
      \item `$pointer` Integer. Index pointing to the selected fleet or index group (e.g., 1 means the first valid fleet group).
    }
  \item `$maturity_weights` List. Settings for weighting maturity-at-age used for spawning stock biomass.
    \itemize{
      \item `$fleet` Logical. Whether to use fleet-specific weights.
      \item `$index` Logical. Whether to use index-specific weights.
      \item `$pointer` Integer. Index pointing to the selected fleet or index group.
    }
}}
}
\value{
List. An updated `em_info` object with:
  \itemize{
    \item Aggregated `catch_info` (`agg_catch`, `catch_paa`).
    \item Aggregated `index_info` (`agg_indices`, `index_paa`).
    \item Updated `par_inputs` with fleet/index aggregations.
    \item Updated weight-at-age (`user_waa`) and `basic_info$waa`.
  }
}
\description{
Aggregates catch and index data from the operating model (`om`) and updates
the estimation model (`em_info`) with the aggregated data. It also adjusts
`par_inputs` and weight-at-age (`user_waa`) to reflect the modified fleet,
region, index, and stock structure.
}
\details{
- Fleets and indices with `fleet_pointer = 0` or `index_pointer = 0` are ignored during aggregation.
- **Aggregated parameters** (`catch_cv`, `index_cv`, `q`, etc.) will return **vectors of length equal to the number of aggregated fleets or indices**.
- If `aggregate_catch_info` or `aggregate_index_info` are missing values, they are replaced with the first element of `em_info$par_inputs`.
- Weight-at-age (`waa`) is aggregated based on fleet or index weights, using either **catch-weighted averages or equal proportions**.
}
