% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_em_input.R
\name{make_em_input}
\alias{make_em_input}
\title{Generate Input Data for the Estimation Model}
\usage{
make_em_input(
  om,
  em_info,
  M_em,
  sel_em,
  NAA_re_em,
  move_em,
  em.opt,
  em_years,
  year.use,
  age_comp_em,
  aggregate_catch_info = NULL,
  aggregate_index_info = NULL,
  filter_indices = NULL,
  reduce_region_info = NULL,
  global_waa = TRUE
)
}
\arguments{
\item{om}{List. The operating model containing observed data and simulation outputs.}

\item{em_info}{List. A set of estimation model parameters used to define the model structure.}

\item{M_em}{List. Natural mortality random effects.}

\item{sel_em}{List. Selectivity random effects.}

\item{NAA_re_em}{List. Numbers-at-age random effects.}

\item{move_em}{List. Movement random effects.}

\item{em.opt}{List. Options for movement in the estimation model.
\itemize{
  \item `$separate.em` Logical. `TRUE` indicates no global SPR, `FALSE` allows global SPR.
  \item `$separate.em.type` Integer (1–3). Specifies assessment model type if `separate.em = TRUE`:
    \itemize{
      \item `1` - Panmictic (spatially aggregated).
      \item `2` - Fleets-as-areas.
      \item `3` - Separate assessment models for each region (`n_regions` models).
    }
  \item `$do.move` Logical. Whether movement is included (if `separate.em = FALSE`).
  \item `$est.move` Logical. Whether movement rates are estimated (if `separate.em = FALSE`).
}}

\item{em_years}{Vector. Years used in the assessment model.}

\item{year.use}{Integer. Number of years used in the assessment model.}

\item{age_comp_em}{Character. Likelihood distribution for age composition data.
\itemize{
  \item `"multinomial"` (default)
  \item `"dir-mult"`
  \item `"dirichlet-miss0"`
  \item `"dirichlet-pool0"`
  \item `"logistic-normal-miss0"`
  \item `"logistic-normal-ar1-miss0"`
  \item `"logistic-normal-pool0"`
  \item `"logistic-normal-01-infl"`
  \item `"logistic-normal-01-infl-2par"`
  \item `"mvtweedie"`
  \item `"dir-mult-linear"`
}}

\item{aggregate_catch_info}{List (optional). User-specified catch aggregation settings 
for panmictic models using aggregate catch.
\itemize{
  \item `$n_fleets` Integer. Number of fleets.
  \item `$catch_cv` Numeric vector (`n_fleets`). Catch CVs for each fleet.
  \item `$catch_Neff` Numeric vector (`n_fleets`). Effective sample sizes for catch.
  \item `$use_agg_catch` Integer vector (`n_fleets`). 0/1 flag for using aggregate catch.
  \item `$use_catch_paa` Integer vector (`n_fleets`). 0/1 flag for using age composition.
  \item `$fleet_pointer` Integer vector (`n_fleets`). Defines fleet grouping (0 = exclude).
  \item `$use_catch_weighted_waa` Logical. Whether to weight-at-age using catch.
}}

\item{aggregate_index_info}{List (optional). User-specified index aggregation settings for panmictic models using aggregate indices.
\itemize{
  \item `$n_indices` Integer. Number of indices.
  \item `$index_cv` Numeric vector (`n_indices`). CVs for survey indices.
  \item `$index_Neff` Numeric vector (`n_indices`). Effective sample sizes for indices.
  \item `$fracyr_indices` Numeric vector (`n_indices`). Fraction of the year for each survey.
  \item `$q` Numeric vector (`n_indices`). Survey catchability coefficients.
  \item `$use_indices` Integer vector (`n_indices`). 0/1 flag for using survey indices.
  \item `$use_index_paa` Integer vector (`n_indices`). 0/1 flag for using age composition.
  \item `$units_indices` Integer vector (`n_indices`). 1 = Biomass, 2 = Numbers.
  \item `$units_index_paa` Integer vector (`n_indices`). 1 = Biomass, 2 = Numbers.
  \item `$index_pointer` Integer vector (`n_indices`). Defines index grouping (0 = exclude).
  \item `$use_index_weighted_waa` Logical. Whether to weight-at-age using indices.
   }}

\item{filter_indices}{Integer (0/1) vector (optional). User-specified which indices are excluded from the assessment model}

\item{reduce_region_info}{List (optional). Specifies modifications for regions. If `NULL`, no modifications are applied.
  The expected components include:
  \itemize{
    \item `$remove_regions`
      Specifies which regions should be removed from the model.

    \item `$reassign`
      Specifies reassignment of surveys from removed regions to non-removed regions.

    \item `$NAA_where`
      Specifies recruitment assignments after region reduction.

    \item `$sel_em`, `$M_em`, `$NAA_re_em`, `$move_em` 
      Model settings that may change based on region reduction.

    \item `$onto_move_list`
      Contains movement-related parameters with the following elements:
      \itemize{
        \item `$onto_move` (array, dimension: `n_stocks × n_regions × (n_regions-1)`)  
          Specifies movement rules between stocks and regions. Default = NULL.

        \item `$onto_move_pars` (array, dimension: `n_stocks × n_regions × (n_regions-1) × 4`)  
          Specifies movement parameters. Default = NULL.

        \item `$age_mu_devs` (array, dimension: `n_stocks × n_regions × (n_regions-1) × n_ages`)  
          Stores age-based movement deviations. If `onto_move == 5`, values are extracted from `basic_info`.
      }
  }}
}
\value{
List. A `wham` input object prepared for stock assessment.
}
\description{
This function generates input data for the estimation model used in 
management strategy evaluation. It constructs the necessary parameters 
and structures for the assessment model.
}
\seealso{
\code{\link{loop_through_fn}}
}
