% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_em_input.R
\name{make_em_input}
\alias{make_em_input}
\title{Generate Input Data for the Estimation Model}
\usage{
make_em_input(
  om,
  em_info,
  M_em,
  sel_em,
  NAA_re_em,
  move_em,
  em.opt,
  em_years,
  year.use,
  age_comp_em,
  aggregate_catch_info = NULL,
  aggregate_index_info = NULL,
  reduce_region_info = NULL,
  filter_indices = NULL,
  global_waa = TRUE
)
}
\arguments{
\item{om}{List. The operating model containing observed data and simulation outputs.}

\item{em_info}{List. A set of estimation model parameters used to define the model structure.}

\item{M_em}{List. Natural mortality random effects.}

\item{sel_em}{List. Selectivity random effects.}

\item{NAA_re_em}{List. Numbers-at-age random effects.}

\item{move_em}{List. Movement random effects.}

\item{em.opt}{List. Options for movement in the estimation model.
\itemize{
  \item `$separate.em` Logical. `TRUE` indicates no global SPR, `FALSE` allows global SPR.
  \item `$separate.em.type` Integer (1–3). Specifies assessment model type if `separate.em = TRUE`:
    \itemize{
      \item `1` - Panmictic (spatially aggregated).
      \item `2` - Fleets-as-areas.
      \item `3` - Separate assessment models for each region (`n_regions` models).
    }
  \item `$do.move` Logical. Whether movement is included (if `separate.em = FALSE`).
  \item `$est.move` Logical. Whether movement rates are estimated (if `separate.em = FALSE`).
}}

\item{em_years}{Vector. Years used in the assessment model.}

\item{year.use}{Integer. Number of years used in the assessment model.}

\item{age_comp_em}{Character. Likelihood distribution for age composition data.
\itemize{
  \item `"multinomial"` (default)
  \item `"dir-mult"`
  \item `"dirichlet-miss0"`
  \item `"dirichlet-pool0"`
  \item `"logistic-normal-miss0"`
  \item `"logistic-normal-ar1-miss0"`
  \item `"logistic-normal-pool0"`
  \item `"logistic-normal-01-infl"`
  \item `"logistic-normal-01-infl-2par"`
  \item `"mvtweedie"`
  \item `"dir-mult-linear"`
}}

\item{aggregate_catch_info}{List (optional). User-specified catch aggregation settings for panmictic models using aggregate catch.
\itemize{
  \item `$n_fleets` Integer. Number of fleets.
  \item `$catch_cv` Numeric vector (`n_fleets`). CVs for annual aggregate catches by fleet.
  \item `$catch_Neff` Numeric vector (`n_fleets`). Effective sample sizes for fleet catches.
  \item `$use_agg_catch` Integer vector (`n_fleets`). 0/1 values flagging whether to use aggregate catches.
  \item `$use_catch_paa` Integer vector (`n_fleets`). 0/1 values flagging whether to use proportions at age observations.
  \item `$fleet_pointer` Integer vector (`n_fleets`). Defines fleet grouping (0 = exclude).
  \item `$use_catch_weighted_waa` Logical. Whether to use weighted weight-at-age based on fleet catches.
}}

\item{aggregate_index_info}{List (optional). User-specified index aggregation settings for panmictic models using aggregate indices.
\itemize{
  \item `$n_indices` Integer. Number of indices.
  \item `$index_cv` Numeric vector (`n_indices`). CVs for annual aggregate index catches.
  \item `$index_Neff` Numeric vector (`n_indices`). Effective sample sizes for survey indices.
  \item `$fracyr_indices` Numeric vector (`n_indices`). Fraction of the year for each survey index.
  \item `$q` Numeric vector (`n_indices`). Initial survey catchabilities.
  \item `$use_indices` Integer vector (`n_indices`). 0/1 values flagging whether to use survey indices.
  \item `$use_index_paa` Integer vector (`n_indices`). 0/1 values flagging whether to use proportions at age observations.
  \item `$units_indices` Integer vector (`n_indices`). 1/2 values flagging whether aggregate observations are biomass (1) or numbers (2).
  \item `$units_index_paa` Integer vector (`n_indices`). 1/2 values flagging whether composition observations are biomass (1) or numbers (2).
  \item `$index_pointer` Integer vector (`n_indices`). Defines index grouping (0 = exclude).
  \item `$use_index_weighted_waa` Logical. Whether to use weighted weight-at-age based on survey catches.
   }}

\item{reduce_region_info}{List (optional). Remove specific regions from the assessment model. If `NULL`, no modifications are applied.
The expected components include:
\itemize{
  \item `$remove_regions` Integer vector (`n_regions`). 0/1 values flagging whether to include regions.
  \item `$reassign` Numeric. Specify reassignment for surveys from the "removed" regions to "remaining" regions.
  \item `$NAA_where`Integer array (`n_stocks × n_regions × n_ages`) 0/1 values flagging whether the stock from a certain age can be present in a specific region.
  \item `$sel_em` Selectivity configuration for the "reduced" assessment model.
  \item `$M_em` Natural mortality configuration for the "reduced" assessment model.
  \item `$NAA_re_em` Numbers-at-age configuration for the "reduced" assessment model.
  \item `$move_em` Movement configuration for the "reduced" assessment model.
  \item `$onto_move_list` List of ontogenetic movement information for the "reduced" assessment model:
    \itemize{
      \item `$onto_move` (array, dimension: `n_stocks × n_regions × (n_regions-1)`)  
        Age-specific movement type.
      \item `$onto_move_pars` (array, dimension: `n_stocks × n_regions × (n_regions-1) × 4`)  
        Parameters controlling age-specific movement patterns.
      \item `$age_mu_devs` (array, dimension: `n_stocks × n_regions × (n_regions-1) × n_ages`)  
        User-specified deviations in mean movement rates across ages.
    }
}}

\item{filter_indices}{Integer (0/1) vector (optional). User-specified which indices are excluded from the assessment model}
}
\value{
List. A `wham` input object prepared for stock assessment.
}
\description{
This function generates input data for the estimation model used in 
management strategy evaluation. It constructs the necessary parameters 
and structures for the assessment model.
}
\seealso{
\code{\link{loop_through_fn}}
}
