% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_em_input.R
\name{make_em_input}
\alias{make_em_input}
\title{Generate input data for the assessment model}
\usage{
make_em_input(
  om,
  em_info,
  M_em,
  sel_em,
  NAA_re_em,
  move_em,
  em.opt,
  em_years,
  year.use,
  age_comp_em = "multinomial"
)
}
\arguments{
\item{om}{Operating model}

\item{em_info}{A list of information used to generate the operating model}

\item{M_em}{Natural mortality random effects}

\item{sel_em}{Selectivity random effects}

\item{NAA_re_em}{Numbers-at-age random effects}

\item{move_em}{Movement random effects}

\item{em.opt}{Movement random effects options
\itemize{
  \item \code{$separate.em} TRUE = No Global SPR, FALSE = Global SPR
  \item \code{$separate.em.type} only if separate.em = TRUE \cr
  {=1} panmictic (spatially-aggregated) \cr
  {=2} fleets-as-areas \cr
  {=3} n single assessment models (n = n_regions) \cr
  \item \code{$do.move} T/F movement is included (use if separate.em = FALSE)
  \item \code{$est.move} T/F movement rate is estimated (use if separate.em = FALSE)
  }}

\item{em_years}{Years used in the assessment model}

\item{year.use}{Number of years used in the assessment model}

\item{age_comp_em}{Likelihood distribution of age composition data in the assessment model
\itemize{
  \item \code{"multinomial"} Default
  \item \code{"dir-mult"}
  \item \code{"dirichlet-miss0"}
  \item \code{"dirichlet-pool0"}
  \item \code{"dir-mult"}
  \item \code{"logistic-normal-miss0"}    
  \item \code{"logistic-normal-ar1-miss0"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl"}
  \item \code{"logistic-normal-pool0"}
  \item \code{"logistic-normal-01-infl-2par"}
  \item \code{"mvtweedie"}
  \item \code{"dir-mult-linear"}
  }}
}
\value{
a wham input
}
\description{
A function to generate the input for the estimation model for management strategy evaluation.
}
\examples{
\dontrun{
data <- generate_basic_info()
input <- prepare_wham_input(basic_info = data)
mod <- fit_wham(input, do.fit = FALSE)
data <- make_em_input(mod, em.opt = list(separate.em = TRUE, separate.em.type = 1, do.move = FALSE, est.move = FALSE), year.use = 10, em_years = 1973:2022)
}
}
\seealso{
\code{\link{loop_through_fn}}
}
