% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mse_output.R
\name{plot_mse_output}
\alias{plot_mse_output}
\title{Generate MSE Output Plots in PNG, HTML, or PDF Format}
\usage{
plot_mse_output(
  mods,
  main_dir,
  output_dir = "Report",
  output_format = c("png", "html", "pdf"),
  width = 10,
  height = 7,
  dpi = 300,
  col.opt = "D",
  method = "mean",
  outlier.opt = NA,
  plot.style = "median_iqr",
  show.whisker = TRUE,
  f.ymin = NULL,
  f.ymax = NULL,
  start.years = 1,
  use.n.years.first = 5,
  use.n.years.last = 5,
  new_model_names = NULL,
  base.model = "Model1"
)
}
\arguments{
\item{mods}{A list of model outputs from multiple simulations and estimation models.}

\item{main_dir}{The main directory path where results should be saved.}

\item{output_dir}{Folder name inside main_dir where plots/reports will be saved (default: "Report").}

\item{output_format}{Output type: "png", "html", or "pdf".}

\item{width}{Width of each plot in inches.}

\item{height}{Height of each plot in inches.}

\item{dpi}{Resolution of saved plots in dots per inch.}

\item{col.opt}{Color palette option for `viridis`-based scales.}

\item{method}{method = NULL: global median, method = "mean": median of means, method = "median": median of medians}

\item{outlier.opt}{outlier.opt = NA (default): no outliers will be shown, outlier.opt = 0 - 25 (26 different shapes to use)}

\item{plot.style}{plot.style = "median_iqr" (Default): only median, IQR will be displayed in the boxplot (no outliers will be displayed), method = "boxplot": boxplot including median, IQR, and outliers}

\item{show.whisker}{Logical, Whether to show 1.5IQR as a whisker}

\item{f.ymin}{Numeric value specifying the lower limit of the y-axis for F-based performance plots (Default = NULL)}

\item{f.ymax}{Numeric value specifying the upper limit of the y-axis for F-based performance plots. (Default = NULL)}

\item{start.years}{The first year used in short-term performance summaries.}

\item{use.n.years.first}{Number of years to summarize at the beginning of the time series.}

\item{use.n.years.last}{Number of years to summarize at the end of the time series.}

\item{new_model_names}{Optional character vector of custom model names.}

\item{base.model}{Model name used as a baseline for relative difference plots.}
}
\value{
Saves output plots/reports to disk; returns nothing explicitly.
}
\description{
This function creates a comprehensive set of plots to visualize MSE (Management Strategy Evaluation)
results from a list of model simulations. Outputs can be saved as individual PNG files or as a single
interactive HTML or PDF report.
}
