% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_F.R
\name{update_om_F}
\alias{update_om_F}
\title{Update Fishing Mortality (F) in the Operating Model}
\usage{
update_om_F(om, year, Fsolve)
}
\arguments{
\item{om}{A fitted operating model that includes burn-in and feedback years.}

\item{year}{Integer. The year for which fishing mortality is being updated.}

\item{Fsolve}{Numeric vector. Estimated fishing mortality values for each fleet.}
}
\value{
An updated operating model (`om`) with:
  \itemize{
    \item Updated F time series.
    \item Adjusted F deviation parameters (\code{F_pars}) for model consistency.
  }
}
\description{
This function updates fishing mortality (F) in the operating model (`om`) for a specified year
and ensures consistency in future F projections.
}
\seealso{
\code{\link{get_F_from_Catch}}, \code{\link{update_om_fn}}
}
