% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_fn.R
\name{update_om_fn}
\alias{update_om_fn}
\title{Update the operating model and generate data}
\usage{
update_om_fn(om, interval.info = NULL, seed = 123, random = "log_NAA")
}
\arguments{
\item{om}{Operating model with years including burn-in + feedback years.}

\item{interval.info}{Catch advice for a number of years projected from the estimation model.
\itemize{
  \item \code{$years} Projection years.
  \item \code{$catch} Matrix (n_region x n_years) projected catch.
}}

\item{seed}{Seed used to generate data.}

\item{random}{A vector of processes that are treated as random effects in the operating model}
}
\value{
An operating model with simulated data and updated F time series.
}
\description{
Function to update F in the operating model (see \code{\link{update_om_F}}) and generate data given the updated F.
}
\seealso{
\code{\link{get_F_from_Catch_region}}
}
