% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_om_fn.R
\name{update_om_fn}
\alias{update_om_fn}
\title{Update the Operating Model and Generate Simulated Data}
\usage{
update_om_fn(
  om,
  interval.info = NULL,
  seed = 123,
  random = "log_NAA",
  method = "nlminb",
  by_fleet = TRUE,
  do.brps = FALSE,
  sim_ecov = TRUE
)
}
\arguments{
\item{om}{A fitted operating model that includes both burn-in and feedback years.}

\item{interval.info}{A list containing projected catch advice for future years from the estimation model. It includes:
\describe{
  \item{\code{years}}{Vector of projection years.}
  \item{\code{catch}}{Matrix (n_regions x n_years) of projected catch.}
}}

\item{seed}{Integer. Seed used for generating simulated data to ensure reproducibility (default = 123).}

\item{random}{Character vector of process names that are treated as random effects in the operating model (default = "log_NAA").}

\item{method}{Character. Optimization method used for solving F. Options include:
\itemize{
  \item \code{"nlminb"} (default): Uses `nlminb` optimization.
  \item \code{"BFGS"}: Uses `optim` with the BFGS method.
}}

\item{by_fleet}{Logical. If TRUE, estimates F separately for each fleet. If FALSE, estimates a single global F (default = FALSE).}

\item{do.brps}{Logical. If TRUE, calculates reference points in the operating model (default = FALSE).}

\item{sim_ecov}{Logical. Whether to generate pseudo ecov data for each realization, only use if ecov is included in the operating model (default = TRUE).}
}
\value{
An updated operating model (`om`) with:
  \itemize{
    \item Updated F time series.
    \item Simulated observation data.
    \item Updated random effect parameters.
  }
}
\description{
This function updates fishing mortality (F) in the operating model, projects the population forward,
and generates new simulated data based on the updated F time series.
}
\seealso{
\code{\link{get_F_from_Catch}}, \code{\link{update_om_F}}
}
